возвращаю float

+++ MySolution/Calculator.cs
@@ -6,9 +6,9 @@ public static class Calculator
 {
     private static readonly char[] Operations = { '-', '+', '*', '/' };
 
-    public static int GetResultExpression(string expression)
+    public static float GetResultExpression(string expression)
     {
-        var stackForOperands = new Stack<int>();
+        var stackForOperands = new Stack<float>();
         var elementsExpression = expression.Split(",", StringSplitOptions.RemoveEmptyEntries);
         var pattern = new Regex(@"\d+", RegexOptions.Compiled);
 
@@ -31,7 +31,7 @@ public static class Calculator
         return result;
     }
 
-    private static int Calculate(int operand1, int operand2, string operation)
+    private static float Calculate(float operand1, float operand2, string operation)
     {
         var res = operation switch
         {
@@ -45,7 +45,7 @@ public static class Calculator
         return res;
     }
 
-    private static void CalculateIntermediateResult(Stack<int> stackForOperands, string element)
+    private static void CalculateIntermediateResult(Stack<float> stackForOperands, string element)
     {
         var b = stackForOperands.Peek();
         stackForOperands.Pop();
